Rename main.cc to pekwm.cc, preparing for more commands.
To avoid confusion with only one main.cc with a main and then more files
with main, rename main.cc to pekwm.cc being the same as the command.
